회원등록기능 분리 / 코드 수정
