+++ src/controller/userController.js
@@ -21,35 +21,8 @@ export class UserController {
     });
     res.status(200).send(user);
   };
-  static createUser = async (req, res) => {
-    const { userPreference, password, ...userFields } = req.body;
-    const received = userPreference ? userPreference.receivedEmail : false;
-    const profileImage = req.file;
-
-    let image;
-    if (profileImage) {
-      image = { create: { url: `/files/user-profiles/${profileImage.filename}` } };
-    } else {
-      image = undefined;
-    }
-
-    const user = await prisma.user.create({
-      data: {
-        ...userFields,
-        password,
-        userPreference: {
-          create: {
-            receivedEmail: received,
-          },
-        },
-        image,
-      },
-    });
-    const { password: _, ...userWithoutPassword } = user;
-    res.status(201).send(userWithoutPassword);
-  };
   static getUserDetail = async (req, res) => {
-    const { id } = parseInt(req.params.id, 10);
+    const id = parseInt(req.params.id, 10);
     const user = await prisma.user.findUniqueOrThrow({
       where: { id },
       select: {
@@ -63,10 +36,10 @@ export class UserController {
     res.status(200).send(user);
   };
   static patchUser = async (req, res) => {
-    const { id } = parseInt(req.params.id, 10);
+    const userId = parseInt(req.params.id, 10);
     const { userPreference, ...userFields } = req.body;
     const user = await prisma.user.update({
-      where: { id },
+      where: { id: userId },
       data: {
         ...userFields,
         userPreference: { update: { receivedEmail: userPreference.receivedEmail } },
@@ -76,10 +49,10 @@ export class UserController {
     res.send(user);
   };
   static deleteUser = async (req, res) => {
-    const { id } = parseInt(req.params.id, 10);
+    const userId = parseInt(req.params.id, 10);
 
     await prisma.user.delete({
-      where: { id },
+      where: { id: userId },
     });
     res.sendStatus(204);
   };
