+++ src/routers/productRouter.js
@@ -20,7 +20,7 @@ productRouter
 //app.use('/files', express.static('uploads')); //>> app.js 미들웨어로 추가
 //상품 상세 조회 , 상품 업데이트 , 상품 삭제
 productRouter
-  .route('/:id')
+  .route('/:productId')
   .get(tryCatchHandler(ProductController.getProductDetail))
   .patch(validate(PatchProduct), tryCatchHandler(ProductController.patchProduct))
   .delete(tryCatchHandler(ProductController.deleteProduct));
